# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba tecnica",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener un listado de projectos",
                "description": "Devuelve un listado de projectos con sus tareas relacionadas",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Projectos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "proyecto nuevo"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "proyecto nuevo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimonio no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Registrar un projecto",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Projecto nuevo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion del proyecto"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projecto registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Testimonio registrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name_customer": [
                                                    "El nombre del cliente es obligatorio"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener un projecto por id",
                "description": "Devuelve un prjecto de projectos con sus tareas relacionadas",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "responses": {
                    "200": {
                        "description": "Projecto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "proyecto nuevo"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "proyecto nuevo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimonio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Actualizar un projecto",
                "operationId": "491d11cc6efd20335421a859159e2487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Projecto nuevo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion del proyecto"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projecto registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Testimonio registrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projecto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projecto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Eliminar un projecto",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID deL projecto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projecto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projecto eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projecto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projecto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "Projects"
        }
    ]
}